Turn off debug mode
We probably shouldn’t be deploying the way that we are; running elm-live
in development and then just uploading the generated index.html to
github and having netlify deploy that document. We accidentally forgot
that we’d been running elm-live with the debugger on and deployed that
copy of the generated index.html. This commit reverts that change.